change 'commentPost' to 'addComment' and several improper codes
